Mantenimiento de Aurora Api Core

+++ app/Http/Controllers/ItemController.php
@@ -5,12 +5,13 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Item;
 use App\Models\Submodule;
+use App\Models\Permission_item;
 
 class ItemController extends Controller
 {
     public function index()
     {
-        $items = Item::all();
+        $items = Item::with(['submodule:id,submodule'])->get();
 
         return response()->json([
             'message' => 'Registros encontrados', 
@@ -21,7 +22,7 @@ class ItemController extends Controller
 
     public function show($id)
     {
-        $item = Item::find($id);
+        $item = Item::with(['submodule:id,submodule'])->find($id);
 
         if (!$item) {
             return response()->json([
@@ -108,6 +109,7 @@ class ItemController extends Controller
         }
 
         $item->item = $request->item;
+        $item->submodule_id = $request->submodule_id;
         $item->path = $request->path;
         $item->description = $request->description;
         $item->icon = $request->icon;
@@ -132,6 +134,24 @@ class ItemController extends Controller
             ], 404);
         }
 
+        $permissions = Permission_item::where('item', $id)
+        ->where(function($query) {
+            $query->where('r', 1)
+                ->orWhere('w', 1)
+                ->orWhere('u', 1)
+                ->orWhere('d', 1);
+        })
+        ->get();
+        
+        if ($permissions->isNotEmpty()) 
+        {
+            return response()->json([
+                'message' => 'El item no se puede eliminar porque tiene permisos asignados',
+                'status' => 400
+            ], 400);
+        }
+
+        Permission_item::where('item', $id)->delete();
         $item->delete();
 
         return response()->json([
@@ -140,4 +160,5 @@ class ItemController extends Controller
             'status' => 200
         ], 200);
     }
+    
 }
